clean-up save page state
