+++ backend/templates/frontend.html
@@ -2,13 +2,8 @@
 <html>
 
 <head>
-    <!-- <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css"> -->
     <link rel="stylesheet" href="{{ url_for('static', filename='css/bootstrap.min.css') }}">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/app.css') }}">
-
-    <!-- <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/4.0.1/socket.io.min.js"></script>
-    <script src="https://cdn.plot.ly/plotly-latest.min.js"></script> -->
 
     <script src="{{ url_for('static', filename='js/jquery-3.5.1.min.js') }}"></script>
     <script src="{{ url_for('static', filename='js/socket.io.min.js') }}"></script>
@@ -23,13 +18,6 @@
         <h1 class="text-center my-5">BWCT Video Analysis Tool</h1>
         <h2>Upload Video</h2>
 
-        <!-- <form id="upload-form" action="/" method="post" enctype="multipart/form-data" class="mt-3">
-            <div class="custom-file">
-                <input type="file" name="video" id="video" class="custom-file-input">
-                <label class="custom-file-label" for="video">Select video</label>
-            </div>
-            <button type="submit" class="btn btn-primary mt-3">Upload Video</button>
-        </form> -->
         <button id="video-upload-button" class="btn btn-secondary mt-3">Open Video</button>
         <button id="raw-data-button" class="btn btn-secondary mt-3">Import Raw Data</button>
 
@@ -38,25 +26,17 @@
                 aria-valuemin="0" aria-valuemax="100"></div>
         </div>
 
+
         <div id="processing-indicator" style="display: none;">
-            <p>Uploading video, please wait. This may take several minutes...</p>
-            <!-- <div class="d-flex flex-row justify-content-center">
-                <div id="loader" class="save_spinner" style="display:none;"></div>
-            </div> -->
+            <p>Importing video, please wait. This may take several minutes...</p>
         </div>
 
 
         <div class="video-container" id="video-container">
             <video id="video-player" controls style="width: 100%; height: 100%;"></video>
             <canvas id="canvas" style="position: absolute; top: 0; left: 0; width: 100%; height: auto;"></canvas>
-            <!-- <div id="loader" class="save_spinner" style="position:absolute ; top: 0; left: 0;display:none;"></div> -->
-
         </div>
 
-        <!-- <form id="upload-day-night-form" method="post" enctype="multipart/form-data" class="mt-3">
-            <input type="file" id="csvFile" name="day_night_file" accept=".csv" />
-            <button type="submit" class="btn btn-primary mt-3">Upload Day Night csv</button>
-        </form> -->
 
         <div class="form-group">
             <label for="save-video">Save Annotated Video</label>
@@ -76,16 +56,13 @@
 
             <button id="download-lines-button" class="btn btn-secondary mt-3" style="display: none">Save Line
                 Crossings</button>
-            <!-- <a id="download-lines-link" href="/download_lines" download style="display: none">Download Line
-                Crossings</a> -->
+
 
             <button id="download-counts-button" class="btn btn-secondary mt-3" style="display: none">Save Counts
                 Output</button>
-            <!-- <a id="download-counts-link" href="" download style="display: none">Download Counts Output</a> -->
 
             <button id="download-processed-video-button" class="btn btn-secondary mt-3" style="display: none">Save
                 Processed Video</button>
-            <!-- <a id="download-processed-video-link" href="" download style="display: none">Download Processed Video</a> -->
 
             <button id="download-plots-button" class="btn btn-secondary mt-3" style="display: none">Save
                 Graphs</button>
@@ -101,7 +78,6 @@
 
 
         <div id="counts-plot"></div>
-        <!-- Add this to your upload.html where you want to display the counts -->
         <div id="counts-display" style="margin-top:   20px;">
             <table id="counts-table" class="table">
                 <thead>
@@ -130,6 +106,6 @@
     </div>
 
 
-    <script src="{{ url_for('static', filename='js/backend.js')}}"></script>
+    <script type="module" src="{{ url_for('static', filename='js/backend.js')}}"></script>
 
 </body>
